ProjectsControler
CRUD operation

+++ PowerUpBuildersWeb/Repositories/ProjectRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using PowerUpBuildersWeb.Models;
 
@@ -7,6 +8,7 @@ namespace PowerUpBuildersWeb.Repositories
     public class ProjectRepository : IProjectRepository
     {
         private readonly AppDbContext _context;
+        
         public ProjectRepository(AppDbContext context)
         {
             _context = context;
@@ -15,7 +17,10 @@ namespace PowerUpBuildersWeb.Repositories
         public void DeleteProject(int projectId)
         {
             Project project = _context.Projects.Find(projectId);
-            _context.Projects.Remove(project);
+            if (project != null)
+            {
+                _context.Projects.Remove(project);
+            }
         }
 
         public Project GetProjectById(int projectId)
@@ -31,6 +36,7 @@ namespace PowerUpBuildersWeb.Repositories
         public void InsertProject(Project project)
         {
             _context.Projects.Add(project);
+            Save();
         }
 
         public void UpdateProject(Project project)
@@ -42,5 +48,7 @@ namespace PowerUpBuildersWeb.Repositories
         {
             _context.SaveChanges();
         }
+
+        
     }
 }
